daemon/requirements.txt
pyzmq~=25.1.1